Rename example app to "LLM Service"; add i18n (13 langs) + flag picker + private save/load
Rebrand the Android example app to a generic, trademark-safe identity:
- App label "LLM Service"; applicationId/namespace/package net.ladenthin.android.llmservice
- Folder llama-android-example/ -> android-llmservice/ (CI job build-android-llmservice,
  run-android-llmservice-test.sh, android-llmservice-aab artifact, gitignore/docs updated)

Add three features:
- Internationalization: every UI string externalized; translations for 13 languages
  (en + de/es/fr/it/pt/ru/tr/ar[RTL]/hi/zh-CN/ja/ko), locales_config.xml.
- In-app flag language picker (top-bar dropdown) via AppCompatDelegate.setApplicationLocales,
  persisted across restarts (AppCompat autoStoreLocales; MainActivity now AppCompatActivity).
  The ViewModel survives the locale-change recreation, so the chat is not lost.
- Private local save/load session: Save/Load buttons persist the conversation + model path
  to filesDir/session.json (org.json, no new dep); readable only by the app, nothing uploaded.

Docs note Gemma 3 4B Instruct as the recommended real on-device model. The instrumented
Compose UI test moves to the new package and still drives the real on-device flow.
