+++ android-llmservice/gradle.properties
