--- llama-android-example/README.md
@@ -1,148 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2026 Bernard Ladenthin <bernard.ladenthin@gmail.com>
-
-SPDX-License-Identifier: MIT
--->
-# llama.cpp Android example app
-
-A minimal, **KISS local-AI** Android chat demo built on
-[`net.ladenthin:llama-android`](../llama-android) (the AAR) and
-[`net.ladenthin:llama-kotlin`](../llama-kotlin) (the coroutines façade). It does exactly
-two things:
-
-1. **Pick a GGUF model** from the device's file system (Storage Access Framework — no
-   storage permission, Google-Play compliant).
-2. **Chat** with it, fully **on-device**, tokens streaming into a Jetpack Compose UI.
-
-Nothing else. It is deliberately one screen. Use an **instruct/chat** GGUF (e.g. a small
-Qwen/Llama/Gemma instruct model) for coherent replies.
-
-> Minimum Android API **28** (Android 9 Pie) — the AAR's floor. Ships native code for
-> `arm64-v8a` (phones) and `x86_64` (emulators / x86 Android).
-
-## Why a `content://` URI is copied to a real path
-
-llama.cpp memory-maps the model from a **real filesystem path**, but the file picker
-returns a `content://` URI. So `ChatViewModel` copies the picked file into the app's
-private `filesDir` and loads it by path. This is the Play-safe route — no
-`MANAGE_EXTERNAL_STORAGE`, no `READ_EXTERNAL_STORAGE`. (Large models take a moment to
-copy; the UI shows a loading state.)
-
-## Build it — no Android Studio required
-
-Android Studio is only an IDE; the build is plain Gradle + AGP. You need a JDK 17 and the
-Android SDK command-line tools (`sdkmanager`) with a platform + build-tools; point Gradle
-at the SDK via `local.properties` (`sdk.dir=/path/to/Android/sdk`) or the
-`ANDROID_HOME`/`ANDROID_SDK_ROOT` env var. Then, from the repo root:
-
-```bash
-# The example resolves the AAR + Kotlin façade from mavenLocal, so publish them first
-# (or depend on a released version from Maven Central and skip these two lines).
-mvn -q -pl llama,llama-kotlin -am -DskipTests -Dgpg.skip=true install
-#   ...plus stage the Android natives + `gradle -p llama-android publishToMavenLocal`
-#   (CI does this automatically — see .github/workflows/publish.yml build-android-example).
-
-VERSION=$(mvn -q -DforceStdout help:evaluate -Dexpression=project.version -pl llama | tail -n1)
-
-# Debug APK (installable on a device/emulator):
-gradle -p llama-android-example assembleDebug -PjllamaVersion="$VERSION"
-
-# Release App Bundle for Google Play (.aab):
-gradle -p llama-android-example bundleRelease -PjllamaVersion="$VERSION"
-```
-
-`-PjllamaVersion` selects which `llama-android`/`llama-kotlin` version to consume; omit it
-to use the default pinned in `app/build.gradle.kts`. Opening the `llama-android-example/`
-folder in Android Studio works too — it's a normal Gradle project.
-
-## Signing & Google Play
-
-**The Maven Central GPG key cannot sign an APK/AAB** — Android signing uses a Java keystore
-(PKCS12/JKS + RSA), a different cryptosystem. For Play, the recommended model is **Play App
-Signing**: you create a one-time **upload keystore**, sign the AAB with it, and Google
-manages the final app-signing key. Create the upload key once:
-
-```bash
-keytool -genkeypair -v -keystore upload-keystore.jks -alias upload \
-        -keyalg RSA -keysize 4096 -validity 10000
-```
-
-`app/build.gradle.kts` reads the keystore + passwords from env vars (or `-P` properties):
-
-| Purpose        | Env var                        | Gradle property             |
-|----------------|--------------------------------|-----------------------------|
-| Keystore path  | `JLLAMA_UPLOAD_STORE_FILE`     | `jllamaUploadStoreFile`     |
-| Store password | `JLLAMA_UPLOAD_STORE_PASSWORD` | `jllamaUploadStorePassword` |
-| Key alias      | `JLLAMA_UPLOAD_KEY_ALIAS`      | `jllamaUploadKeyAlias`      |
-| Key password   | `JLLAMA_UPLOAD_KEY_PASSWORD`   | `jllamaUploadKeyPassword`   |
-
-When no keystore is supplied, the release build falls back to **debug signing** so local
-builds, forks, and PR CI still produce an installable artifact without any secret.
-
-```bash
-JLLAMA_UPLOAD_STORE_FILE=$PWD/upload-keystore.jks \
-JLLAMA_UPLOAD_STORE_PASSWORD=… JLLAMA_UPLOAD_KEY_ALIAS=upload JLLAMA_UPLOAD_KEY_PASSWORD=… \
-  gradle -p llama-android-example bundleRelease -PjllamaVersion="$VERSION"
-# -> app/build/outputs/bundle/release/app-release.aab  (upload to the Play Console)
-```
-
-The first upload of a new package to the Play Console must be done manually (to register the
-app + enroll in Play App Signing). After that, automated Play uploads are possible with the
-Gradle Play Publisher plugin or fastlane `supply` (needs a Play service-account JSON) — left
-out here to keep the demo KISS.
-
-## CI: build, sign, and a **real** on-device test
-
-The `build-android-example` job in `.github/workflows/publish.yml`:
-
-1. Builds the core jar + installs `llama-kotlin` to mavenLocal, stages the CI-built Android
-   natives, and publishes the `llama-android` AAR to mavenLocal.
-2. Builds a **release `.aab`** — signed with the upload key when the CI secrets are set
-   (`ANDROID_UPLOAD_KEYSTORE_BASE64` = base64 of the keystore, plus the store/alias/key
-   password secrets), otherwise debug-signed — and uploads it as the `llama-android-example-aab`
-   artifact.
-3. Runs the app's **instrumented UI test on a KVM-accelerated x86_64 emulator** (API 30, the
-   same runner as `test-android-emulator`): it adb-pushes the tiny cached draft model, then
-   `connectedDebugAndroidTest` launches the app, **types a prompt, taps Send, and asserts a
-   non-empty assistant reply streamed back** — exercising the Compose UI, the `ChatViewModel`,
-   the `generateChatFlow` façade, and real native inference from `jni/x86_64/libjllama.so`
-   together.
-
-### Why an instrumented (emulator) test, not Espresso-in-a-vacuum
-
-The valuable test here is the **on-device** one — Espresso / Compose UI tests only mean
-something when run on a real device or emulator (`connectedAndroidTest`), which is exactly
-what the CI job does. This app uses a **Compose UI test** (`createEmptyComposeRule` +
-`ActivityScenario`) rather than classic Espresso, but they're interchangeable for driving the
-UI. The one thing deliberately **not** UI-tested is the system file-picker dialog: it runs in
-a separate process (only `UiAutomator` could drive it, and it's flaky), so the test injects the
-model path via an `Intent` extra and drives everything after the picker. See
-`app/src/androidTest/.../ChatFlowInstrumentedTest.kt`.
-
-Run the same test locally against a booted emulator/device:
-
-```bash
-adb push <tiny-model>.gguf /data/local/tmp/jllama-test-model.gguf
-adb shell chmod 644 /data/local/tmp/jllama-test-model.gguf
-gradle -p llama-android-example connectedDebugAndroidTest -PjllamaVersion="$VERSION"
-# The test self-skips (JUnit Assume) if the model file is absent.
-```
-
-## Layout
-
-```
-llama-android-example/
-├── settings.gradle.kts          # AGP 8.7.3 + Kotlin 2.4.0 + Compose plugin, mavenLocal first
-├── gradle.properties
-└── app/
-    ├── build.gradle.kts         # minSdk 28, Compose, release signingConfig (upload key/debug fallback)
-    ├── proguard-rules.pro       # stub — the AAR ships the JNI keep rules
-    └── src/
-        ├── main/AndroidManifest.xml
-        ├── main/res/values/{strings,themes}.xml
-        ├── main/kotlin/net/ladenthin/llama/example/
-        │   ├── MainActivity.kt      # Compose UI + SAF model picker
-        │   └── ChatViewModel.kt     # model load + streaming chat (the only logic)
-        └── androidTest/kotlin/net/ladenthin/llama/example/
-            └── ChatFlowInstrumentedTest.kt   # end-to-end on-device UI test
-```
