--- llama-android-example/app/src/main/AndroidManifest.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-SPDX-FileCopyrightText: 2026 Bernard Ladenthin <bernard.ladenthin@gmail.com>
-
-SPDX-License-Identifier: MIT
--->
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <!--
-      No storage permissions: the model is picked with the Storage Access Framework
-      (ACTION_OPEN_DOCUMENT), which grants one-shot read access to the chosen file with
-      no permission prompt and is fully Google-Play compliant (no MANAGE_EXTERNAL_STORAGE).
-    -->
-    <application
-        android:allowBackup="true"
-        android:label="@string/app_name"
-        android:supportsRtl="true"
-        android:theme="@style/Theme.LlamaExample">
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-    </application>
-</manifest>
