--- llama-android-example/app/src/main/res/values/strings.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-SPDX-FileCopyrightText: 2026 Bernard Ladenthin <bernard.ladenthin@gmail.com>
-
-SPDX-License-Identifier: MIT
--->
-<resources>
-    <string name="app_name">llama.cpp Android chat</string>
-</resources>
