--- llama-android-example/app/src/main/res/values/themes.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-SPDX-FileCopyrightText: 2026 Bernard Ladenthin <bernard.ladenthin@gmail.com>
-
-SPDX-License-Identifier: MIT
--->
-<resources>
-    <!--
-      A bare platform theme (no AppCompat / Material dependency needed): Compose supplies
-      its own Material 3 theming in MainActivity. NoActionBar because the Compose UI draws
-      its own top bar.
-    -->
-    <style name="Theme.LlamaExample" parent="@android:style/Theme.Material.Light.NoActionBar" />
-</resources>
